Destroy random obstacles & %bonus spawn chance
When an obstacle moves past the players untouched it is deleted, and there is a chance that a random live obstacle is destroyed. The obstacle spawn amount per tile is now also a float, and the fractional part of the float is a % chance to round up. (2.75f gives 75% chance for 3, 25% for 2)